made code styling changes
