+++ Eatery Blue/UI/EateryCard/EateryExpandableCardContentView.swift
@@ -119,6 +119,7 @@ class EateryExpandableCardContentView: UIView {
     @objc private func didTapEateryDetails(_ sender: UITapGestureRecognizer) {
         if let navigationController = findNavigationController() {
             guard let eatery = eatery else { return }
+            
             let eateryVC = EateryModelController()
             eateryVC.setUp(eatery: eatery)
             navigationController.pushViewController(eateryVC, animated: true)
