+++ Eatery Blue/UI/EateryCard/EateryExpandableCardDetailView.swift
@@ -131,6 +131,7 @@ class EateryExpandableCardDetailView: UIView {
     @objc private func didTapEateryDetails(_ sender: UITapGestureRecognizer) {
         if let navigationController = findNavigationController() {
             guard let eatery = eatery else { return }
+            
             let eateryVC = EateryModelController()
             eateryVC.setUp(eatery: eatery)
             navigationController.pushViewController(eateryVC, animated: true)
